Rework block handling
Make the piece of state for the block being built a `Maybe` instead of
detecting empty blocks and not emitting those.  We start out a function
with a `Nothing` block.  When starting a new block we can just not emit
the previous block if it's `Nothing`. When emitting an instruction and
the current block is `Nothing`, we implicitly start a new block with a
fresh unnamed name.  This will typically only happen at the start of a
function if no block is started, and should match the way LLVM
handles blocks at the the start of functions.