Add emitBlockStart function
... and make it possible to create recursive blocks without using
`RecursiveDo`.

`emitBlockStart` is a low-level function that emits a block using a
given name (assumed to be fresh), which means that the user can manually
use `freshName` to create a block name that can be used before the block
is emitted.

The reason for not always wanting `RecursiveDo` is that it's sometimes
difficult or unclear how to make your code lazy enough for it to be able
to work, which can lead to crashes and infinite loops that are difficult
to diagnose.