Avoid constant folding during instruction encoding
The LLVM IRBuilder API that we use for encoding instructions does a
limited form of constant folding. This causes problems for us since
we expect that the result of encoding an instruction is an Instruction
and not a Constant (e.g., we try to set metadata which isn’t possible
for a Constant). In the best case this leads to an assertion failure,
in the worst case it will just segfault.