Handle NamedTypeReference in extractValue
extractValue would error if given an alias to a struct/array,
since it relied on extractValueType that doesn't know about
aliases.

Now it calls out the same function gep uses for type resolution.
Because gep also accepts pointers and vector types, we check
beforehand that only the the expected types were given

Fixes #295