Fix typeOf const struct whose type is named
For a constant structure `c`, `typeOf c` would always return
a structure type with the type of all the values. However,
when the constant structure has a named type reference as
type, that should be used instead.

This issue manifests when using llvm-hs-pretty, that can then
produce invalid LLVM IR. For example, if we read the following
code:

```
%t = type { i8 }
@g = global { i32, %t } {i32 42, %t {i8 42}}
```

When pretty-printed we get:

```
%t = type { i8 }
@g = global { i32, %t } {i32 42, {i8} {i8 42}}
```

which LLVM rejects due to the `%t` vs `{i8}` mismatch.