Remove dumpModule and dumpMetadata functions
The symbols they call in the LLVM C bindings are not resolved (as of 12 Apr 2021 and possibly before). (#322)

It seems reasonable to expect users of llvm-hs to be able to compose `writeFile stderr <$> ppLlvm <$> moduleAST`

Since LLVM 12 changed the mask argument of `shufflevector` to a constant vector of `Int32`, `ppLlvm` requires a `Typed` instance for `[Int32]` for the above snippet to work properly.