IR-builder: use first terminator during codegen
Previously, the latest terminator instruction was always used (`br`,
`condBr`, `ret`, ...). This limited composability when creating
combinators based on the IR builder.

Now, we keep track of the first terminator which is more in line with
how LLVM semantics work (terminators that appear later in a basic block
are dead code).

With this change, we can now create control flow combinators that work
like we expect them to:

```haskell
-- A combinator for a "one sided if statement"
if' condition asm = mdo
  condBr condition begin end
  begin <- block `named` "if.begin"
  asm
  br end
  end <- block `named` "if.end"
  return ()

example = do
  function "f" [(AST.i1, "a"), (AST.i1, "b")] AST.i1 $ \[a, b] -> mdo

    entry <- block `named` "entry"; do
      if' a $ do
        if' b $ do
          -- Previously, this ret instruction would not be emitted!
          ret (bit 0)

      ret (bit 1)
```