Remove legacy PassManager, add limited support for new PassBuilder
LLVM is in the process of removing the legacy PassManager APIs that
llvm-hs depends on pretty heavily. To harden against future breakage
this commit adds preliminary support for the new PassBuilder-based
infrastructure. However, since it is quite different than the previous
API, only curated pass specs are supported for now.