fix: GHC 9.0 deep subsumption
Fix the problems introduced by https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0287-simplify-subsumption.rst

Roughly, `g` is not always equal to `(\x -> g x)`.

In GHC 9.2.4 and GHC 9.4, this was reverted and a new extension
`DeepSubsumption` was introduced to change the behavior, with the old
behavior by default. However it is unclear if this will be backported to
GHC 9.0.3 and at least, using the new syntax fix the build for 9.0.X and
9.2.X < 9.2.4.