Fix compilation error with OrcJITC.cpp
Problem: With LLVM 15.0.7 on my machine, `llvm-hs` does not compile the
C++ module at `llvm-hs/src/LLVM/Internal/FFI/OrcJITC.cpp`. This is
because `InProcessMemoryManager` now expects a page size as its
argument.

Furthermore, the import of `LLVM.AST.Type` is redudant so I removed it.

Solution: There is now `InProcessMemoryManager::Create` that tries to
automatically detect the page size, so we can use it to resolve the
problem. Also, delete the redundant import.